pyproject.toml
[project]
name = "docuzebre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.38.0",
    "pydantic>=2.8.2",
    "kor>=2.0.0",
    "langchain-openai>=0.1.23",
    "pymupdf>=1.24.10",
    "langchain>=0.2.15",
    "langchain-core>=0.2.37",
    "pytest",
    "python-dotenv"
]
[tool.pytest.ini_options]
pythonpath = "front"
testpaths = "tests"
addopts = "--maxfail=2 -rf"  # exit after 2 failures, report fail info